Crate private_event_sourcing_integrity

Structs

Acknowledgement

AcknowledgementContent

EventHistory

EventSentToRecipients

EventSentToRecipientsContent

Message

PrivateEventContent

PrivateEventEntry

ReceiveMessageInput

SignedContent

SignedEntry

Enums

AwaitingDependencies

EntryTypes

UnitEntryTypes

Functions

__num_entry_types

action_hash

entry_defs

genesis_self_check

Validation you perform during the genesis process. Nobody else on the network performs it, only you. There is no access to network calls in this callback

validate

This is the unified validation callback for all entries and link types in this integrity zome Below is a match template for all of the variants of DHT Ops and entry and link types

validate_agent_joining

Validation the network performs when you try to join, you can’t perform this validation yourself as you are not a member yet. There is access to network calls in this function

validate_create_acknowledgement

validate_create_awaiting_dependencies

validate_create_event_history

validate_create_events_sent_to_recipients

validate_create_private_event

validate_delete_acknowledgement

validate_delete_awaiting_dependencies

validate_delete_event_history

validate_delete_events_sent_to_recipients

validate_delete_private_event

validate_update_acknowledgement

validate_update_awaiting_dependencies

validate_update_event_history

validate_update_events_sent_to_recipients

validate_update_private_event

Type Aliases

SignedEvent