Crate private_event_sourcing_integrity

Structs

AwaitingDependencies

EncryptedMessage

PrivateEventEntry

SignedContent

SignedEvent

Enums

EntryTypes

LinkTypes

UnitEntryTypes

Functions

__num_entry_types

__num_link_types

action_hash

entry_defs

genesis_self_check

Validation you perform during the genesis process. Nobody else on the network performs it, only you. There is no access to network calls in this callback

validate

This is the unified validation callback for all entries and link types in this integrity zome Below is a match template for all of the variants of DHT Ops and entry and link types

validate_agent_joining

Validation the network performs when you try to join, you can’t perform this validation yourself as you are not a member yet. There is access to network calls in this function

validate_create_awaiting_dependencies

validate_create_encrypted_message

validate_create_link_agent_encrypted_message

validate_create_private_event

validate_delete_awaiting_dependencies

validate_delete_encrypted_message

validate_delete_link_agent_encrypted_message

validate_delete_private_event

validate_update_awaiting_dependencies

validate_update_encrypted_message

validate_update_private_event